Fix Node/State ID data type inconsistency: Convert uint32_t to uint16_t across Automaton framework
Changes:
- Edge_Node.h: Node IDs and Edge destination IDs changed to uint16_t (matches AP ID sizing)
- Automaton.h: nodeMap key and all method parameters changed to uint16_t
- BuchiAutomaton.h/cpp: acceptingStates, initialState, and all state ID parameters changed to uint16_t
- TS.h/cpp: initialStates vector and all state ID parameters changed to uint16_t; stateToNodeId map keys changed to uint16_t
- ProductAutomaton.h/cpp: stateMapping pairs, acceptingStates, and all state ID parameters changed to uint16_t
- TS.cpp: Constructor cast operations updated for uint16_t state IDs

Rationale: Node IDs map to atomic propositions (~65K max), so uint16_t is appropriate. Edge weights/counts remain uint32_t to allow more edges than nodes. This reduces memory footprint and improves type consistency.

+++ Automatons/Automaton.h
@@ -20,7 +20,7 @@ class Automaton{
     protected:
         uint32_t numNodes = 0;
         uint32_t numEdges = 0;
-        std::map<uint32_t, Node*> nodeMap;  //map from node ID to node pointer for quick access
+        std::map<uint16_t, Node*> nodeMap;  //map from node ID to node pointer for quick access
     
     public:
         Automaton() = default;
@@ -34,10 +34,10 @@ class Automaton{
         virtual void add_Node(Node* node) = 0;
         
         //check if there is a direct edge between two nodes
-        virtual bool isAdjacent(uint32_t srcId, uint32_t dstId) const = 0;
+        virtual bool isAdjacent(uint16_t srcId, uint16_t dstId) const = 0;
         
         //get a node by ID
-        Node* getNode(uint32_t nodeId) const {
+        Node* getNode(uint16_t nodeId) const {
             auto it = nodeMap.find(nodeId);
             if (it != nodeMap.end()) {
                 return it->second;
@@ -46,7 +46,7 @@ class Automaton{
         }
         
         //get all nodes
-        const std::map<uint32_t, Node*>& getNodes() const {
+        const std::map<uint16_t, Node*>& getNodes() const {
             return nodeMap;
         }
 };
